requirements.txt
flask
twilio
python-dotenv
openai
pymongo
sentence-transformers
numpy
